pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "africastalking>=2.0.2",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.1",
    "joblib>=1.5.3",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "numpy>=2.4.3",
    "ollama>=0.6.1",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "reportlab>=4.4.10",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "scrapy>=2.14.2",
    "selenium>=4.41.0",
    "streamlit>=1.55.0",
    "sqlmodel>=0.0.22",
    "sqlalchemy>=2.0.35",
    "asyncpg>=0.29.0",
    "pyjwt>=2.8.0",
    "aiocache>=0.12.2",
    "uvicorn>=0.42.0",
    "webdriver-manager>=4.0.2",
    "psycopg2-binary>=2.9.11",
]
